fixing pretty hard to find memory bug.
In a bunch of places we used references to capture stuff and
then afterwards modified the underlying vector storage, pretty fun that
i made this error so many times and there are probably a bunch more
of them in the codebase.

Weirdly i never ran into segfaults on my own machine, but macos with
clang instanstly gave invalid output (still no segfault)

+++ compiler/compiler/lifeness.cpp
@@ -68,11 +68,11 @@ Lifetimes lifeness(const CanoModel &model) {
       continue;
     auto n = g.get(memory::NodeId(i));
     for (const auto &e : n->outgoing()) {
-      const std::size_t d = static_cast<std::size_t>(e.dst().id());
+      const std::uint64_t d = static_cast<std::uint64_t>(e.dst().id());
       if (!active[d])
         continue;
       for (const auto &s : e.srcs()) {
-        const std::size_t sid = static_cast<std::size_t>(s.id());
+        const std::uint64_t sid = static_cast<std::uint64_t>(s.id());
         if (!active[sid])
           continue;
         succ[sid].push_back(d);
